refactored header script into generic style script; other tests not implemented yet
